# Conflicts:
#	src/Illuminate/Foundation/Application.php
#	src/Illuminate/Support/Facades/DB.php
#	src/Illuminate/View/Engines/CompilerEngine.php

+++ src/Illuminate/Pagination/AbstractPaginator.php
@@ -481,7 +481,7 @@ abstract class AbstractPaginator implements Htmlable
     public static function resolveCurrentPage($pageName = 'page', $default = 1)
     {
         if (isset(static::$currentPageResolver)) {
-            return call_user_func(static::$currentPageResolver, $pageName);
+            return (int) call_user_func(static::$currentPageResolver, $pageName);
         }
 
         return $default;
